Refactor: 사용자에게 입력받는 기능 클래스로 분리
-UserInput.js에 숫자 입력받는 기능과 재시작 입력받는 기능 정리
